# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "School API",
        "description": "API for managing school data",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api"
        }
    ],
    "paths": {
        "/api/v1/sessions": {
            "get": {
                "tags": [
                    "school-v1.1",
                    "school-v1.7",
                    "school-v1.8",
                    "school-v1.9",
                    "school-v2.0"
                ],
                "summary": "Get list of sessions",
                "description": "Returns list of sessions",
                "operationId": "getSessionsList",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "school-v1.1"
                ],
                "summary": "Store new session",
                "description": "Returns session data",
                "operationId": "storeSession",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "2025/2026"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-09-01"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-07-31"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/v1/sessions/{id}": {
            "get": {
                "tags": [
                    "school-v1.1"
                ],
                "summary": "Get session information",
                "description": "Returns session data",
                "operationId": "getSessionById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Session id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": 123
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "put": {
                "tags": [
                    "school-v1.1"
                ],
                "summary": "Update existing session",
                "description": "Returns updated session data",
                "operationId": "updateSession",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Session id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "start_date",
                                    "end_date"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "2025/2026"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "example": "2025/2026"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-09-01"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-07-31"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "delete": {
                "tags": [
                    "school-v1.1"
                ],
                "summary": "Delete existing session",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteSession",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Session id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/v1/sessions/{id}/terms": {
            "get": {
                "tags": [
                    "school-v1.1",
                    "school-v1.7",
                    "school-v1.8",
                    "school-v1.9",
                    "school-v2.0"
                ],
                "summary": "Get list of terms for a session",
                "description": "Returns list of terms for a session",
                "operationId": "getTermsForSession",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Session id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "school-v1.1"
                ],
                "summary": "Store new term",
                "description": "Returns term data",
                "operationId": "storeTerm",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Session id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "1st"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-09-01"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-07-31"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/v1/terms/{id}": {
            "get": {
                "tags": [
                    "school-v1.1",
                    "school-v1.7",
                    "school-v1.8",
                    "school-v1.9",
                    "school-v2.0"
                ],
                "summary": "Get term information",
                "description": "Returns term data",
                "operationId": "getTermById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Term id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "school-v1.1"
                ],
                "summary": "Update existing term",
                "description": "Returns updated term data",
                "operationId": "updateTerm",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Term id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "2nd"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "example": "2nd"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-09-01"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-07-31"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "delete": {
                "tags": [
                    "school-v1.1"
                ],
                "summary": "Delete existing term",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteTerm",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Term id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}